chore: moved quiz components files
