+++ src/components/features/credit-score/credit-score.tsx
@@ -12,7 +12,7 @@ import { AvatarFile } from '@/components/atoms/avatar-file';
 import { LoadingButton } from '@/components/atoms/buttons/loading-button';
 import { ShareButton } from '@/components/atoms/buttons/share-button';
 import { ReadMore } from '@/components/atoms/read-more-less';
-import { TokenFilled } from '@/components/molecules/mint-card/assets/token-filled';
+import { TokenFilled } from '@/components/organisms/mint/mint-card/assets/token-filled';
 import { HolderDialog } from '@/components/organisms/holder-dialog';
 import { ClientNav } from '@/components/organisms/navbar/client-nav';
 import { ROUTES } from '@/constants/routes';
