+++ src/components/features/gates/create/tasks/quiz-task/components/question-creator.tsx
@@ -2,14 +2,15 @@ import {
   CreateGateData,
   Question,
 } from '@/components/features/gates/create/schema';
-import { QuestionField } from '@/components/features/gates/create/tasks/quiz-task/question-field/question-field';
-import { RadioCheckBoxCreator } from '@/components/molecules/radio-checkbox-creator/radio-checkbox-creator';
+import { QuestionField } from '@/components/features/gates/create/tasks/quiz-task/components/question-field';
 import { DragDropContext, Droppable, Draggable } from 'react-beautiful-dnd';
 import { useFormContext } from 'react-hook-form';
 
 import CloseIcon from '@mui/icons-material/Close';
 import DragIndicatorIcon from '@mui/icons-material/DragIndicator';
 import { Divider, IconButton, Stack } from '@mui/material';
+
+import { RadioCheckBoxCreator } from './radio-checkbox-creator';
 
 export function QuestionCreator({
   questions,
