+++ src/components/organisms/mint/mint-card/assets/gateway-gray.tsx
