+++ src/components/organisms/mint/mint-card/assets/token-filled.tsx
