+++ src/components/organisms/mint/mint-card/index.ts
