+++ src/components/organisms/mint/mint-card/mint-card.tsx
