+++ src/components/organisms/mint/mint-card/process.tsx
