+++ src/components/organisms/mint/mint-card/screen/default-mint-screen.tsx
