+++ src/components/organisms/mint/mint-card/screen/start-mint-screen.tsx
