+++ src/components/organisms/mint/mint-card/utlis/animate-message.tsx
