+++ src/components/organisms/mint/mint-card/utlis/categories.tsx
